LPS-135823 Refactor DB Partition tests
